Introduce block time protocol V2 (15 second windows)
- Port and adapt PIVX commits on time protocol V2 (PR #1002)
- Port PIVX commits on nTime offset warnings (PR #1138)
- Adjust chain parameters
- Add several copyright notices

+++ src/miner.h
@@ -162,7 +162,7 @@ public:
 
     /** Construct a new block template with coinbase to scriptPubKeyIn */
     std::unique_ptr<CBlockTemplate> CreateNewBlock(const CScript& scriptPubKeyIn,
-            std::shared_ptr<CMutableTransaction> pCoinstakeTx = nullptr, std::shared_ptr<CStakeInput> coinstakeInput = nullptr, unsigned int nTxNewTime = 0);
+            std::shared_ptr<CMutableTransaction> pCoinstakeTx = nullptr, std::shared_ptr<CStakeInput> coinstakeInput = nullptr, uint64_t nTxNewTime = 0);
 
 private:
     // utility functions
